+++ src/main/java/com/meoguri/linkocean/controller/profile/dto/CreateProfileRequest.java
@@ -5,7 +5,7 @@ import static java.util.stream.Collectors.*;
 import java.util.List;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
-import com.meoguri.linkocean.domain.profile.service.dto.RegisterProfileCommand;
+import com.meoguri.linkocean.domain.profile.service.command.dto.RegisterProfileCommand;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
