+++ src/main/java/com/meoguri/linkocean/controller/profile/dto/GetProfilesResponse.java
@@ -1,6 +1,6 @@
 package com.meoguri.linkocean.controller.profile.dto;
 
-import com.meoguri.linkocean.domain.profile.service.dto.GetProfilesResult;
+import com.meoguri.linkocean.domain.profile.service.command.dto.GetProfilesResult;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
