+++ src/main/java/com/meoguri/linkocean/domain/profile/entity/FavoriteCategories.java
@@ -23,6 +23,7 @@ import lombok.NoArgsConstructor;
 public class FavoriteCategories {
 
 	private static final int MAX_FAVORITE_CATEGORIES_SIZE = Category.values().length;
+
 	@ElementCollection
 	@CollectionTable(name = "favorite_category")
 	@Column(name = "category")
