+++ src/main/java/com/meoguri/linkocean/domain/profile/service/command/FollowService.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.profile.service;
+package com.meoguri.linkocean.domain.profile.service.command;
 
 public interface FollowService {
