+++ src/main/java/com/meoguri/linkocean/domain/profile/service/command/dto/GetProfileTagsResult.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.profile.service.dto;
+package com.meoguri.linkocean.domain.profile.service.command.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
