+++ src/main/java/com/meoguri/linkocean/domain/profile/service/command/dto/UpdateProfileCommand.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.profile.service.dto;
+package com.meoguri.linkocean.domain.profile.service.command.dto;
 
 import java.util.List;
 
@@ -9,14 +9,11 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public final class GetDetailedProfileResult {
+public final class UpdateProfileCommand {
 
 	private final long profileId;
 	private final String username;
 	private final String image;
 	private final String bio;
-	private final List<Category> favoriteCategories;
-	private final boolean isFollow;
-	private final int followerCount;
-	private final int followeeCount;
+	private final List<Category> categories;
 }
