+++ src/main/java/com/meoguri/linkocean/domain/profile/service/query/ProfileQueryService.java
@@ -1,0 +1,5 @@
+package com.meoguri.linkocean.domain.profile.service.query;
+
+public interface ProfileQueryService {
+	
+}
