+++ src/main/java/com/meoguri/linkocean/domain/profile/service/query/ProfileQueryServiceImpl.java
@@ -1,0 +1,10 @@
+package com.meoguri.linkocean.domain.profile.service.query;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional(readOnly = true)
+@Service
+public class ProfileQueryServiceImpl implements ProfileQueryService {
+	
+}
