+++ src/test/java/com/meoguri/linkocean/domain/bookmark/service/TagServiceImplTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.meoguri.linkocean.domain.profile.service.dto.GetProfileTagsResult;
+import com.meoguri.linkocean.domain.profile.service.command.dto.GetProfileTagsResult;
 import com.meoguri.linkocean.test.support.service.BaseServiceTest;
 
 class TagServiceImplTest extends BaseServiceTest {
